+++ helpers/pass-rest.js
@@ -1,3 +1,3 @@
 import _ from 'lodash';
 
-export const passRest = props => _.omit(props, 'lfFramework', 'lfComponent', 'lfLocale', 'transformer');
+export const passRest = props => _.omit(props, 'lfFramework', 'lfComponent', 'lfLocale', 'lfOnEnter', 'transformer');